Refactor codecs
